Support python3.
These changes allow the unit tests to run successfully using python3.4,
without affecting 2.7 support.  I'm not 100% sure if it is enough for
full python3 support since then only thing I've run with it is the tests,
and I'm not clear how complete they are.

Only one change is to the library code, the rest of the changes just
generalize the syntax used in the tests to syntax that is supported by
both python2 and python3.  The library change adds support for python3
string filenames without changing the python2 filename support.